修复了一个bug

+++ study_credit_course.py
@@ -218,7 +218,7 @@ def end_study():
 if __name__ == "__main__":
     print("课程学分：" + str(credit_list))
     select_credit = input("请输入要学习的课程的学分：")
-    if int(select_credit) not in credit_list:
+    if float(select_credit) not in credit_list:
         print("输入错误。告辞")
         exit(0)
     else:
